requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
